alterado

+++ conversor_de_moedas.py
@@ -1,7 +1,7 @@
 
 
 
-dolar = int(input('Digite o valor em dolar para conversão:'))
+dolar = float(input('Digite o valor em dolar para conversão:'))
 
 real= dolar*5.68
